pyproject.toml
[project]
name = "translator-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.69.0",
    "python-dotenv>=1.1.1",
]
